Introduce BaseDBModel, UUIDMixin, TimestampMixin to eliminate field duplication
- Create app/models.py with ORM mixins and AI SDK message schemas
- Collapse app/models/ package into single app/models.py file
- BaseDBModel (UUID PK + timestamps): used by 11 models
- TimestampMixin only: ThreadDB (string PK)
- UUIDMixin only: PersonalAccessTokenDB (no updated_at column)
- Use sa_column_kwargs instead of sa_column to avoid shared Column instances
