+++ backend/app/agents/mcp_servers/repository.py
@@ -4,7 +4,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlmodel import select
 
 from app.agents.models import AgentMCPServerDB
-from app.repositories import BaseRepository
+from app.repository import BaseRepository
 
 
 class AgentMCPServerRepository(BaseRepository[AgentMCPServerDB]):
