--- backend/app/repositories/__init__.py
@@ -1,3 +1,0 @@
-from app.repositories.base import BaseRepository
-
-__all__ = ["BaseRepository"]
